app/requirements.txt
appnope==0.1.2
backcall==0.2.0
bidict==0.21.2
blinker==1.4
blis==0.7.4
Brotli==1.0.9
brotlipy==0.7.0
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chromedriver-py==90.0.4430.24
click==7.1.2
cryptography==3.4.7
cymem==2.0.5
decorator==5.0.3
dnspython==1.16.0
emoji==0.6.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.0.0/en_core_web_trf-3.0.0-py3-none-any.whl
eventlet==0.30.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SocketIO==5.0.1
get-docker-secret==1.0.1
gevent==21.1.2
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httplib2==0.19.1
hyperframe==6.0.1
idna==2.10
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
kaitaistruct==0.9
MarkupSafe==1.1.1
murmurhash==1.0.5
mysql-connector-python==8.0.23
nltk==3.6.2
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
parso==0.8.2
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
Pygments==2.8.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.16.0
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
selenium==3.141.0
selenium-wire==4.3.0
six==1.15.0
smart-open==3.0.0
spacy==3.0.5
spacy-alignments==0.8.3
spacy-legacy==3.0.2
spacy-transformers==1.0.2
spacytextblob==3.0.1
srsly==2.4.0
textblob==0.15.3
thinc==8.0.2
tokenizers==0.10.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==1.0.1
wsproto==1.0.0
zope.event==4.5.0
zope.interface==5.4.0
